Rename NavigationMock dir
