--- lib/src/Mock.ts
@@ -1,1 +1,0 @@
-export * from './Mock';
