+++ lib/src/Mock/NativeEventsReceiver.ts
